Don't fail hard if Symbols for Device could not be found
Summary:
The symbols within `--sysroot` are only needed for symbolication. No need to fail hard if the user hasn't downloaded them yet.

This will be improved in time such that we will add pulling down Symbols from a device. For the time being, it's up to the user to do this in advance via Xcode

Reviewed By: jbardini

Differential Revision: D33161361

fbshipit-source-id: 92b7331a772d578c73eb79500f928ed15346e5a4