Use direct read/writes in separate queues for debugserver wiring
Summary:
For whatever reason, it appears that using an `FBFileReader` on the socket fd does not result in bytes coming through in a timely manner. I suspect this is due to some underlying buffering that I don't quite understand in the bowels of `dispatch_io`.

However, it's actually very easy to have a read and write loop operate in two background queues. In this case we're using `-[NSFIleHandler availableData]` which does produce all data, as little as a single byte, whenever it is ready. This means that `lldb` will no longer hang due to receiving partial data over the protocol

Reviewed By: RuijieC-dev

Differential Revision: D33477041

fbshipit-source-id: 7d35e06dc7c6f63e56b713064fb73075bc878eca