dsym install: keep untar directory if no bundle ID is specified
Summary:
when installing dsym via idb there are two ways to specify the source:

1. a single path in a local filesystem
2. URL pointing to an archive

Now 1) works as expected. In case of 2) however the archive might contain a few dsym's and that breaks the expectation introduced in D32111894 (https://github.com/facebook/idb/commit/903034bc961a50576ea4a444be62202e7c3455bb), that there is only one dSYM in the archive. If more than one dSYM have been added to the archive the best action we could do is to treat the entire directory as a dSYM

Reviewed By: lawrencelomax

Differential Revision: D33508172

fbshipit-source-id: 415ea59cb4b63d0d379f292a281c5a9060aef853