Add FBContainedFile file wrapper
Summary:
This abstraction is going to be very important for using path mappings against non host-filesystem containers (for example applying path mappings to AFC on devices).

This is done by shifting responsibility for low-level file operations to a protocol, which can then be adopted elsewhere. This protocol can then be provided in path mapping containers. This layer of abstraction can then be used for devices. For the time being, this change is a no-op. The intention is that this lays the groundwork for later changes to devices.

Additionally, this fixes a bug (for which there is now a unit test) where pushing to the root of a mapped container would succeed (but fail silently)

Reviewed By: jamescoggan

Differential Revision: D32945493

fbshipit-source-id: e9c40788dd711c8f6e28af66ffe2b946f21e03b8