Sort Disk Image Listing
Summary:
Lists disk images in the correct order, so that it's easier to eyeball in output. Adds a compare method so that this can be repeated later on.

Internally the `--disk-images` container uses a dictionary, so this means that ordering goes out of the window. This means when constructing the listing we need to run the sort on the values.

Reviewed By: jamescoggan

Differential Revision: D33581623

fbshipit-source-id: ac9e237a22477b8bd7d502b923eefb52d8bbf2e4