Add shared cache extraction in symbol container
Summary:
Now that we have the ability to extract symbols rather than just a shared cache, we can expose this in the file commands. This is done through a symbols "Directory", which is really just an alias that we can later pull.

This will go through the process as described in D33579655 and then place the results in this directory.

Reviewed By: jamescoggan

Differential Revision: D33579658

fbshipit-source-id: c0053cf2f01fe1f1f835859254e335c829644c0a