Add FBProvidedFile Conformance to FBAFCConnection
Summary:
This puts `FBAFCConnection` behind the same abstraction used for host level files (`FBProvidedFile`). This now means that `FBAFCConnection` can now be wrapped in a `FBFileContainer` the same as host-level files. This removes one area of implementation divergence between sims and devices, even though the behaviour is identical.

There's not a huge amount of difference to the implementation as a whole now. This just requires moving of methods so that they are implemented in the right place and confirms to the `FBFileContainer` itself. There's additional implementations here for fileExists and for pulling. This pulling implementation now supports pulling a dir, but that won't be supported until D33427968, where we swap out the `FBFileContainer` wrapper to use the same one as sims.

Reviewed By: jamescoggan

Differential Revision: D33426326

fbshipit-source-id: 75e0b87695abad108a10578a52a11e27e77c254c