Incremental writes of data AFC Push
Summary:
Currently, we pull the entire contents of the host-resident file into memory before pushing it in parts to the device.

Now we use an input stream to read the host file and incrementally push it to the device. This can be done with a fixed and re-used buffer to avoid excessive allocations. Large files will now avoid allocating the entire memory.

Reviewed By: jamescoggan

Differential Revision: D33567100

fbshipit-source-id: 5dbaaf7ffdaeb73899f9051eb5cbd1f5de645953