Remove NSBundle load swizzle
Summary:
This was needed when we were indirectly loading Xcode Plugins.

Now we no longer do this, Xcode plugins are not loaded directly or indirectly. As a result we no longer need to surpress this. This was needed in the early days of `fbsimctl` where Xcode plugins would load Swift runtime overlays.

Since we no longer load these plugins, we don't have to worry about such conflicts

Reviewed By: jamescoggan

Differential Revision: D34038989

fbshipit-source-id: 5dbdd62447c3c40f9a5896004db41cb2100467ad