Read and apply avg_bitrate parameter
Summary: This change will apply avg_bitrate parameter passed through GRPC video_stream API. The value will be then passed through "options" to the encoder as a kVTCompressionPropertyKey_AverageBitRate property. Related field which determines maximum bitrate will be automatically deduced as 1.5 that of the average bitrate. In the future we might allow to pass that as well

Reviewed By: jamescoggan

Differential Revision: D34213757

fbshipit-source-id: 6873880f5e786d285ffcf33a5790e32009cac7bc