fix broken logic for contacts and camera permissions
Summary:
In iOS13 and newer permissions for camera and contacts were broken. Until this diff these permissions were grant sending a message to the simulator device object. This is an object from a private framework. Ultimately it would result into changes in the TCC.db. However, in newer versions of iOS simulator device no longer responds to such messages. So this fix pivots to modifying the TCC.db directly.
Incidentally, apple's own simctl cli fails to grant permission for contacts correctly. Possibly this is related to the same issue.

Reviewed By: jbardini

Differential Revision: D34343347

fbshipit-source-id: 051100d65e09784ba36342bc84972554f09c099e