docs: add social banner in support of Ukraine (#766)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156676751-9d70061d-d0a3-4859-a09c-ad8302c270d0.png)

Pull Request resolved: https://github.com/facebook/idb/pull/766

Reviewed By: jbardini

Differential Revision: D34630503

Pulled By: dmitryvinn-fb

fbshipit-source-id: b0f4f8c44b5aa080846216728c5b2c5823105da3