(Un)install progress callback, error message of last event
Summary:
The logging for the install/uninstall is very rich internally. However, we could do a better job propagating this in the error message.

In order to facilitate that, I'm adding an object that contains this progress and does the logging. This is then referenced from install/uninstall as the callback context. The last event is copied and then kept around so that if an error occurs, this last event can be read back and put in the error message

Reviewed By: jamescoggan

Differential Revision: D34167579

fbshipit-source-id: 5aa34c603e0471152296d35a54822f2dec1fe80c