Move grpc method handlers to distinct entities
Summary:
grpc-swift chose the way to have single huge protocol with all method definition of service. On the long run we will again have enormous `CompanionServiceProvider` class with 2-3k LOC
Intention of this diff is to implement each command in distinct `MethodHandler` to have more maintainability and separation of concerns.

Reviewed By: lawrencelomax

Differential Revision: D34518958

fbshipit-source-id: e3084b31a7ea32ee1aaa291de05adbb970f814f4