Disconnect simulator services upon shutdown
Summary:
If a simulator dies or is shut down behind the back of the process using `FBSimulatorControl`, the hid/framebuffer will stay around and subsequently fail in confusing ways.

We should explicitly handle this case, such that we can be sure that if the simulator was to boot again for the current idb process, we would be able to handle this new boot through the restoration of hid/framebuffer alongside the new booted simulator.

Reviewed By: jbardini, Nekitosss

Differential Revision: D34752002

fbshipit-source-id: c1a3f6dace2495657c71ea6ce0cb60e1c7f6b14d