Add swift port on startup
Summary:
Adding support for swift grpc port specification and setting the `IDB_SWIFT_COMPANION_PORT` env var for companion launches if it's set.
The new port is validated: errors are silently ignored (but logged) but if the port bind was successful then we return a new env var to the client.

Reviewed By: Nekitosss

Differential Revision: D34615016

fbshipit-source-id: 35185b1e2a9326afc6498cdcb3097d975e6160cc