Follow updated private headers in SimDisplayIOSurfaceRenderable
Summary:
Private headers was changed in new Xcode. We need to follow this accordingly.

The underlying implementation is `ROCKRemoteProxy` and requesting the objects actually performs call to some service. The selector object may and may not exists, so checks like `[surface respondsToSelector: selector(ioSurface)]` may return true, because this selector exists in proxy implementation, but underlying object may still not implement that selector and we will receive nil result.

To be 100% sure we calling both methods (with plural underlying implementation and not) and one of the implementation will succeed.

I also was thinking about should we split now `SimDisplayIOSurfaceRenderable` into `Legacy` and `New`, but decided to just add new methods and use single protocol, because it was implemented this was in `SimulatorKit.framework` for Xcode 13.2 and it is better to implement it similarly to it than to split implementation

Differential Revision: D35547591

fbshipit-source-id: 8cf93ffe2c26677c0a260e528acb2a4247ff6cc2