Return `id` from `-[FBIDBCommandExecutor accessibility_info_at_point:nestedFormat:]`
Summary:
This method returns a NSDictionary or a NSArray, depending on `point` being nil or not. This is missleading and causes a crash when bridging the return value to Swift.

The callees of `accessibility_info_at_point:nestedFormat:` only json-serialise the accessibility info returned, so we can use `id` to 'correctly' represent both return types. Until we refactor this API to something explicitly, and correctly, typed.

Reviewed By: Nekitosss

Differential Revision: D35841517

fbshipit-source-id: 170c242f423a0794b46941d49ab7a0b9cb2e66fb