Fix futures chaining in `runTestWithLaunchConfiguration:reporter:logger:`
Summary:
`+[FBXcodeBuildOperation confirmExitOfXcodebuildOperation:configuration:reporter:target:logger:]` returns a Future, so we want to `fmap` it.

Using `map` meant the chain was not waiting for the completion of the future returned by `confirmExitOfXcodebuildOperation:` and possibly introducing a race condition.

Using `map` also meant that the return of `runTestWithLaunchConfiguration:reporter:logger:` is of type `FBFuture<FBFuture<_> *> *` and it crashes in `BridgeFuture.value` since it has the wrong type.

Reviewed By: Nekitosss

Differential Revision: D36515948

fbshipit-source-id: 6d10b9038576dde031d0945f47aa685b570a1a61