Fix undefined behaviour causing crash by preventing closing stdin file descriptor
Summary:
This fixes a bug when `fileDescriptor` can be equal `0` due to empty initial value or `null` descriptor object.
First, closing the `0` file descriptor is not good thing in general (`0` is `stdin` of current process). This is not what we want to do.
Second, this seems like causes undefined behaviour and file descriptors of *grpc-swift* writes damaged and we receive fatal error in `SwiftNIO` library.

Reviewed By: jbardini

Differential Revision: D36543122

fbshipit-source-id: f5deddf9d84643a11cf3860a4b18fbcc34db047e