Show mounted disk image if it's not known to xcode
Summary: We are hidding disk image if we cannot map signature to the list of known disk images. This cause problems: idb thinks there is no disk image mounted and tries to mount it, causing an exception. We cannot remove existing disk image, since it's invisible. I'm adding ability to show a mounted disk image which only signature available as `unknown.dmg` image. This allows deleting it to unblock development on iPhones.

Reviewed By: lawrencelomax, Nekitosss

Differential Revision: D37272807

fbshipit-source-id: 7c99e2d85b259f24966959d80a69a8e32ceb3576