Add in new command to revoke permissions
Summary: This adds in the ability to reset permissions as a complement to `approve`. This naming mirrors the similar `adb` command `pm revoke` which is the opposite of `pm grant`. While this more closely mirrors `xcrun simctl privacy reset` than `xcrun simctl privacy revoke`, `reset` is a little too vague to be used as a top-level `idb` command, so here we are.

Reviewed By: Nekitosss

Differential Revision: D37928629

fbshipit-source-id: 4495baf39163209a3bd927a892c60e701ee3e9f8