Fix grpc-swift source files generation in OSS release builds
Summary:
There was an issue that broke the release build compilation of the OSS xcodeproject.
Default way to create a build rule in Xcode leads to flag that configures multiple build step invocations. That confuses Xcode build system, because we produce the same files regardless of the architecture we call that step against
We do not need multiple calls for each architecture for that build rule, we need to run that command only once

Differential Revision: D38535115

fbshipit-source-id: d137c3df2a53f9c09328c218a55765f97c5b1f8f