Switch to universal binary building for OSS
Summary:
`idb_companion` homebrew build is broken. We moved from cocoapods to Xcode-managed SwiftPM for grpc-swift and this set up do not work inside the homebrew (homebrew creates sandbox + xcode package resolver creates sandbox, that does not supported properly in macOS, so Xcode fails to downlaod the packages).
To fix that we need to move from building from source to binary-only dependencies. That is pretty easy to do.

Moving back to cocoapods is annoying, because `grpc-swift` library is only distributed by SwiftPM

So, we need to have properly built *fat* release binary that will work on every architecture that we will just distribute via Homebrew.

Reviewed By: lawrencelomax

Differential Revision: D38651117

fbshipit-source-id: ca3489f0c38ac1f87c2d50da31f8df4a83de1350