Fixed IDB Crash on Empty Crash String
Summary:
Recently when our test bundle crashes we've stopped getting any useful information out of it. For example: https://www.internalfb.com/intern/testinfra/diagnostics/2533274971659098.281475024247480.1686212626

Looking at the logs, idb is crashing on this line. So most likely we've got an empty crash string. https://www.internalfb.com/intern/everpaste/?handle=GHhShAQy6qXXouoAAOnYVTen-pI6bswMAAAz

This handles the empty string case here so should stop idb crashing, at least on this line. But it's not clear to me if this will improve the overall result or not if the crash string is nil anyway.

Reviewed By: Nekitosss

Differential Revision: D46554618

fbshipit-source-id: bfbf342b7a46ba92344baa6cccdc36943223d6b8