More FBCrashLog from File Error Messaging
Summary:
I'm trying to figure out what's going on with these crashes we keep getting that have no crash log anywhere. For example https://www.internalfb.com/intern/everpaste/?handle=GPNY1wi_QgkFiMECAIpjSKiRYDsjbswMAAAz

It looks like it's been given a crash log to open at .idb_companion-2023-06-16-112130.ips but for whatever reason it can't extract anything from it.

Maybe it never existed in the first place or maybe it exists but it's not yet got anything in it.

To help with debugging it I've broken down the stages of trying to read the file a bit more with more precise error messaaging.

Reviewed By: Nekitosss

Differential Revision: D46813456

fbshipit-source-id: f8d5e3a82aa36c1fe416d55fd5db23b48c1c8f42