Fixing Crash Log Reporting
Summary: I added more logging info in D46813456 but made a mistake. That error I'm checking for the existence of is always returning true. That's because we've got a pointer to an error and I'm checking whether the pointer is nil or not rather than whether the error itself is nil. I've removed the whole check for error thing as if there's a data there's no error anyway.

Reviewed By: jbardini, Nekitosss

Differential Revision: D47056988

fbshipit-source-id: 2921a7f7f3c9cc3f779d6dd2d239f8395909ab6a