Update installation.mdx
When I ran `brew install idb-companion` , I was asked to install Xcode. Which is pretty serious business (in terms of waiting times) so I am adding this to the install instructions.

```shell
% brew install idb-companion
idb-companion: A full installation of Xcode.app 13.0 is required to compile
this software. Installing just the Command Line Tools is not sufficient.

Xcode can be installed from the App Store.
Error: idb-companion: An unsatisfied requirement failed this build.
```