No B-frames for idb video streams
Summary:
This setting doesn't really make sense to me and diff which added doesn't provide context: D34375320. B-frames are a way of telling the decoder that the current frame's data is "in the future." This makes sense when you're showing a seekable video (like one that's fully downloaded), or when you can tolerate some latency (like when downloading a video from a streaming website, you see buffering). It doesn't make sense when you're looking at a live a stream since the future isn't here yet.

This setting either adds ~20 frames of latency, or doesn't do anything. Removing it.

Reviewed By: jbardini

Differential Revision: D48195348

fbshipit-source-id: 8d83ee543b604eeffe6a100ee42d509bfae60891