Add an additional touch down event at the end of swipes to avoid inertial scroll
Summary:
Inertial scroll was added to arm simulators in Xcode 14. This "feature" means that swipes will continue past the swipe end point while slowly decelerating. Apple just trying to keep us on our toes.

This diff adds an additional tap event at the end of the swipe. This tap event is in the same place as the last position of the swipe, and causes the scrolling to "slow down."

Differential Revision: D48929892

fbshipit-source-id: 2f55375c011bcedb3850a7bf23eb4dcf1ea7f5d5