Increase crash wait timeout
Summary:
I went through 100+ idb_companion logs and not once did we find the crash log file. I got access to the sim tier to search for the ips files myself. Most of the time they’re generated. But every crash log that was generated is generated outside of the 1 minute window idb is waiting for. I found about 7 out of 10. The fastest I’ve seen has a file creation date of 1:02 after the crash, the slowest was 14 mins. Locally they show up instantly so it’s weird they’re ridiculously slow on the CI; I’m guessing the thing making them is on a very low priority queue.

This increases the wait timeout to 2 minutes from 1 minute.
Judging from what I've seen I think the distribution will be something along the lines of:
0% collected at 1 minute
20% collected at 2 minutes
50% collected at 5 minutes
75% collected at 15+ minutes.

So we could argue 5 minutes would be better. That's a long time to wait though.

Reviewed By: Nekitosss

Differential Revision: D48520154

fbshipit-source-id: e33c24f303975e8f96c9bfe8e46a4e9cf0c39740