Signal process under test termination uppon crashes
Summary:
Most of test results reporting assume that the test under process is terminated when a crash happens, except coverage collection.

idb only reads coverage collection after `processUnderTestExitedMutable` is resolved to ensure that we get complete coverage data when the tests run successfully. However, the Reporter callback that resolves `processUnderTestExitedMutable` was never called when the test process crashed and the test ran request would hang.

Change `FBLogicReporterAdapter` to `[self.reporter processUnderTestDidExit]` like it does when the test process finishes cleanly. This is safe because `didCrashDuringTest:` is only called when the test process IO is closed, which happen when the process terminates.

Reviewed By: Nekitosss

Differential Revision: D52796488

fbshipit-source-id: a0b249defe9c133e6cf8567304cb8818d068eedc