Fix crasher on describe-point
Summary:
I've found another crasher that occurs, this is when using `describe-point` when there are no elements that match. The best way of reproing this is hitting a point with a fullscreen dialog

{F1456920875}

In this case I am hitting 100, 100. Which is somewhere near the top left of the screen. A full modal will prevent any accessibility requests to anything other than the dialog.

describe-point does not remediate on zero rect. This is because empty/no information is a valid state for this call, but not for describe-all.

Reviewed By: jamescoggan

Differential Revision: D53807689

fbshipit-source-id: 8d92f6d9a2d4481382a521b601615321d7d467c3