Extend the timeout of test listing
Summary:
Recently we found many of our tests are failed with "Failed to list tests." error message.
https://www.internalfb.com/sandcastle/workflow/1292533093057807088/insights?selectedLines=420-463-1-1
https://www.internalfb.com/intern/testinfra/testconsole/testrun/1407375244792239

With one of the affected test `//xplat/arfx/skylight/app/ui/simulator_preferences:SkylightSimulatorPreferencesTestsAppleMac`

After comparing the log output and investigations, early findings are
- this only happens to x86_64 Mac tests, but not to arm64 Mac tests
- affected tests are quite big in binary size (the `xctest` bundle can be 2GB and more)
- the timeout happens at this line
https://www.internalfb.com/code/fbsource/[b10994f3b0de]/fbobjc/EndToEndTests/FBSimulatorControl/src/CompanionLib/FBIDBCommandExecutor.m?lines=291
- the log when the timeout happens are
{P1200928392}
```
list_tests_in_bundle failed with: Waited 60.000000 seconds for process 20037 to terminate, but the xctest process stalled:
```

From the above log, we can see the "Querying Bundle" started but not yet to reach the Principle class line
https://www.internalfb.com/code/fbsource/[5ff9cd430747]/fbobjc/EndToEndTests/FBSimulatorControl/src/Shims/Shimulator/TestReporterShim/XCTestReporterShim.m?lines=546-547,593,705-706,717-718

To repro, a simple executable is made as
{P1201284452}
and then with Instrument to profile - we can see the x86_64 version of the xctest takes much longer to load at the `dlopen` line.
 {F1474118714}
 {F1474118862}

We need more time to investigate why the test bundle is so big and affecting the dlopen loading performance. But to unblock these dying tests, a quick workaround is to extend this timeout.

Reviewed By: jbardini

Differential Revision: D55436652

fbshipit-source-id: ec2f7af29b05f3ffedec958801d456a34920446e