Fix installing test bundles from streams
Summary:
This diff fixes S437790 which blocks debugging unit tests in iOS ODs.

The fixed bug happens when an .xctest bundle is installed by idb_companion through a stream. idb_companion decompresses the archive sent through the stream to a temporary directory, then creates a symlink to the bundle in the temp directory to install it in the simulator. Before actually running the test, the temporary directory is cleaned up, leaving the symlink pointing at nothing.

This diff will *move* the test bundle from the temp directory to install it if the bundle comes from an archive, and will use symlinks for all other cases.

Differential Revision: D60975841

fbshipit-source-id: 342768bb5a34d092128b3c9ab30c5bee05329aba