Fix test running when test host is running
Summary:
When running tests locally in VS Code it's easy for a user to disconnect the debugger without quitting the test host app, and because the simulator is headless, this leaves the test host running with no way to quit it. Because the test host is running, future test runs will always fail because the process is "already running."

This diff changes test runs to kill and relaunch the test host if it is already running, so a user can continue to run tests.

Differential Revision: D61735347

fbshipit-source-id: 1f6230daf2d89b1f85da4f1089e54f04eeb0ecab