Add logging to debug failed listing due to invalid format
Summary: Many tests are failing due to the error "Failed to list tests". Logs include the error "The data couldn’t be read because it isn’t in the correct format." This diff adds logs to see when a test should be added to the listing, and if the listing fails, what the description of the data that should contain the test listing entails.

Reviewed By: lawrencelomax

Differential Revision: D64563764

fbshipit-source-id: e2d75d9935730b205a8d73371ac04c692d832ed9