More logging to debug failed listing due to invalid format
Summary: Many tests are failing due to the error "Failed to list tests". Logs include the error "The data couldn’t be read because it isn’t in the correct format." Current logs indicate that reading the data may be truncated- length of the data seems to be either 0 or smaller than on successful runs. This diff logs the contents of the temporary file that should contain test data if it is not read successfully.

Differential Revision: D65140844

fbshipit-source-id: 23a56b4ec869aa39b46f8cc10ddc39d0f9755c95