Updating run command to dynamically create bundles based on file path (#871)
Summary:
Pull Request resolved: https://github.com/facebook/idb/pull/871

## Context

The installation step for XCTest bundles creates symlinks to the test bundle source on the host machine in a static directory (`idb-test-bundles`), or copies the files from the source location to that directory. This is not necessary for XCTest if the bundle lives in the host machine, as the test bundle is injected as a dylib.

By removing these symlinks, we get rid of some additional complexity, and remove the risk of leaving broken symlinks around on the host device. These old/broken symlinks create noise in test logs, which make them hard to read.

## This diff
- Enables a test request API which takes a `testPath`, which points directly to the original test bundle location.
- Creates a bundle & artifcat objects in memory on run, as opposed to finding them on disk in `idb-test-bundles`.

## Future Considerations
- Possible enable code signing in the test run creation request. The install step, by default, code signs the test bundles.
- We could create a `CommandExecutor` that is not dependent on an instance of the `StorageManager` class. The initialization of the `StorageManager` creates a static directory that we no longer need to exist if we skip the installation step. This is a pretty big undertaking, and this diff removes the symlinks which have been creating

Differential Revision: D66895643

fbshipit-source-id: 1d5dd7499f44d5f5d721a6479d872765707b97ad