Fix test debugging when test host is running
Summary:
When debugging app tests locally it's possible to end a debugging session without actually stopping the test host app. This causes any subsequent test debug sessions to fail because during test setup the test host is "already running".

This diff changes the behavior so that during test debugging setup, if the test host app is already running, it will be relaunched. This fixes the cases where once the test host is running, debugging becomes impossible.

Differential Revision: D67115830

fbshipit-source-id: c5d0bb0e80e4e7cfa590d61ad6823830388aa268