Adding testPath APIs for App and UI XCTests (#872)
Summary:
Pull Request resolved: https://github.com/facebook/idb/pull/872

Follow up to https://github.com/facebook/idb/pull/871.

This PR creates "filePath" APIs for XCTestRequests in order to bypass the need for the test bundle install step. See https://github.com/facebook/idb/pull/871 for a description of the problem.

TL;DR creating symlinks of the test bundle and not cleaning this up creates a bunch of issues where IDB can be pointing to a non-longer-existing testrun or bundle file, leading to failing tests.

Differential Revision: D67093466

fbshipit-source-id: 9158524da3fb1217738749a562f1d512ecca6ebe