Adding file path APIs for "list"
Summary:
## Context

`idb list` produces an abundance of "Bundle not found at path" errors, as it attempts to read test descriptors from a static symlink directory that is not automatically cleaned. This leads to messy error logs, as old symlinks (that have had their source file removed) cannot be loaded properly.

## This diff
- Adds file path APIs for `list` so that we load the descriptors from the source file, as opposed to the symlink proxy.

Differential Revision: D67464085