Use char switch case for FBLoggingWrapper
Summary:
All of the cases here are single-character strings. Therefore they can be switch-cased instead of through repeated `NSString` comparisons. This will significantly reduce the overhead of this function.

A switch over a scalar value (a char) is constant time over the number of cases. A series of if statements is `O(n)` over the cases.

Reviewed By: xgerrit

Differential Revision: D76828816

fbshipit-source-id: c3b9b0989ddb211327c3dd7357f660daf53d2c26