Logging for Simulators uses the UDID as the logger name
Summary: This has been the case for ages, but logging lines frequently contain `simulator_set` as the logger that is placed in the inflated device. The reason for this is that the logging forwarder captures the non-udid-named logger. We just need to use the same logger as that that the `FBSimulator` itself stores.

Reviewed By: xgerrit

Differential Revision: D82806622

fbshipit-source-id: 8fce029ca43078c83df04441a846b890450d8a59