Symlinked Developer Directory
Summary:
There's no need to shell out to `xcode-select -p` since this basically just reads `/var/db/xcode_select_link`.

Instead we can just resolve this path directly and confirm that it is correct in the way that we do from the prior implementation. This avoids a shell-out, which matters for shorter-lived process latency

Reviewed By: xgerrit

Differential Revision:
D82948214

Privacy Context Container: L1246528

fbshipit-source-id: 9301ecf0efb759adf48fdd7718666e56cbe86930