Add FBAXKeys and lazy key filtering to accessibilityDictionaryForElement
Summary:
Fetching properties that aren't used is wasteful. This stack adds APIs to enable that.

Added a public NS_STRING_ENUM (`FBAXKeys`) that defines all 15 accessibility dictionary keys currently being fetched.

This is then used to prevent overfetching of properties within the set. As of now this will be a no-op with future changes threading through the keyset. This diff is a no-op otherwise.

Reviewed By: xgerrit

Differential Revision: D86300417

fbshipit-source-id: 4af4fb658436bde23889e376ac40113d52ae1ee1