Add ability to clear photo library within idb companion
Summary:
SimulatorFrameworkBridge can be used to clear the photo library. This is similar to the capability added to clear contacts.

This uses the Photos framework to access the photo library and clear it. The `kTCCServicePhotos` entitlement is needed for this to work even though it clears photos using private APIs.

Private APIs needed to be used for this as the Photos framework will always show a destructive alert when we try to delete photos/videos from the device.

Differential Revision:
D87447670

Privacy Context Container: L1246528

fbshipit-source-id: 15b9bd13795abfcacf55be3c176a93d12a2820c2