Fix -Wswitch-default issues
Summary:
**Context:** https://fburl.com/switch-enum

----

Now that `-Wswitch-enum` is an error in all of `fbobjc`, making it such that all enum values need a case within a switch regardless of having a default case, we want to also enable `-Wswitch-default` as an error (ensuring we cover all values outside the given enum's values).  This will reduce SEVs as it will eliminate undefined behavior when a value outside the enum range is switched upon.

Differential Revision: D88039760

fbshipit-source-id: 4ed9a12e276f4014b24a5649aad3b18fbf5e73da